Windows package: clean up meson build directories upon exit

+++ windows/build.sh
@@ -32,13 +32,14 @@ ROOT_DIRECTORY=$(dirname "$(readlink -f "$0")")
 [[ -d /tmp/windeps ]] && { echo "Please remove the /tmp/windeps directory" 2>&1; exit 1; }
 [[ -d "$ROOT_DIRECTORY"/../build-win-matlab ]] && { echo "Please remove the build-win-matlab directory" 2>&1; exit 1; }
 [[ -d "$ROOT_DIRECTORY"/../build-win-octave ]] && { echo "Please remove the build-win-octave directory" 2>&1; exit 1; }
+[[ -d "$ROOT_DIRECTORY"/../build-doc ]] && { echo "Please remove the build-doc directory" 2>&1; exit 1; }
 
 # Create temporary folder and make sure it is deleted upon exit (along with /tmp/windeps)
 TMP_DIRECTORY=$(mktemp -d)
 cleanup()
 {
     [[ -z $TMP_DIRECTORY ]] || rm -rf -- "$TMP_DIRECTORY"
-    rm -rf /tmp/windeps
+    rm -rf /tmp/windeps "$ROOT_DIRECTORY"/../build-win-matlab "$ROOT_DIRECTORY"/../build-win-octave "$ROOT_DIRECTORY"/../build-doc
 }
 trap cleanup EXIT
